Refactor code and fix naming issues in test cases
